Adversarial attacks for MNIST model 2
This is the attack notebook for MNIST model 2 (same structure and optimization method, the only difference is that we don't normalize the input to [-1,1]).
We can see that the success rate of all the attacks increase >10%, which is a significant change (even deducting the 5% for the gap in test accuracy, it is still >5%). This very simple experiment implies that even a simple normalization can help increase the robustness of the CNN model (at least for MNIST dataset).